--- src/components/About.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function About() {
-  return (
-    <div>About</div>
-  );
-}
-
-export default About;
