+++ src/components/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Home.css';
+import '../styles/Home.css';
 
 function Home() {
   return (
