+++ src/styles/Home.css
@@ -41,7 +41,7 @@
 
 @media screen and (min-width: 800px) {
   .home__container header {
-    background-image: url(./images/calc.svg);
+    background-image: url(../images/calc.svg);
     background-position: 100% -50%;
     background-repeat: no-repeat;
     padding: 2em 4em;
